pyproject.toml
[project]
name = "wechat-agent"
version = "0.1.0"
description = "白羊 · 微信数字分身 SaaS · AI 副驾驶辅助回复"
requires-python = ">=3.11"
authors = [
    {name = "连大哥", email = "ach73680@gmail.com"},
    {name = "童虎", email = "tonghu@anthropic.local"},
]

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "sqlalchemy>=2.0",
    "greenlet>=3.0",
    "aiosqlite>=0.20",
    "alembic>=1.13",
    "aiohttp>=3.10",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "structlog>=24.4",
    "pyyaml>=6.0",
    "rich>=13.9",
    "apscheduler>=3.10",
    "cryptography>=42.0",
]

[project.optional-dependencies.rag]
sentence-transformers = ">=3.0"
numpy = ">=1.26"

[project.optional-dependencies]
postgres = [
    "asyncpg>=0.30",
    "pgvector>=0.3",
]

llm = [
    # 直接调 LLM API 的官方 SDK（可选 · aiohttp 已能用 · 装 SDK 提速 + 支持流式）
    "openai>=1.50",          # 兼容 DeepSeek/GLM/vLLM 的 OpenAI 协议接口
    "anthropic>=0.39",       # Claude Sonnet 4.6
    "zhipuai>=2.1",          # 智谱 GLM 官方 SDK
]

windows = [
    "wxautox>=39.0.0",
    "humancursor>=1.1",
    "pyqt6>=6.7",
    "pywin32>=308",
]

train = [
    "torch>=2.5",
    "transformers>=4.45",
    "peft>=0.13",
    "trl>=0.11",
    "datasets>=3.0",
    "accelerate>=1.0",
    "bitsandbytes>=0.44",
]

dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "mypy>=1.13",
]

[project.scripts]
baiyang-server = "server.main:run"
baiyang-init-db = "scripts.init_db:main"
baiyang-seed = "scripts.seed_tenant:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["server", "client", "pipeline", "shared"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
